requirements.txt
# Core Framework
flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# LangChain & RAG
langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.3

# Vector Store
chromadb==0.4.22

# Embeddings (local, free)
sentence-transformers>=2.6.0

# Document Processing
markdown==3.5.1
beautifulsoup4==4.12.2

# HTTP & Utils
requests==2.31.0
pydantic==2.5.3

# Evaluation
numpy==1.26.3
tqdm==4.66.1

# Testing (optional)
pytest==7.4.4
